+++ code/game/g_arenas.c
@@ -92,7 +92,7 @@ void UpdateTournamentInfo( void ) {
 		n = level.sortedClients[i];
 		Com_sprintf( buf, sizeof(buf), " %i %i %i", n, level.clients[n].ps.persistant[PERS_RANK], level.clients[n].ps.persistant[PERS_SCORE] );
 		buflen = strlen( buf );
-		if( msglen + buflen + 1 >= sizeof(msg) ) {
+		if( msglen + buflen + 1 >= (int)sizeof(msg) ) {
 			break;
 		}
 		strcat( msg, buf );
